* 'GP3127_quadtree_index_spike' of https://github.com/tst-dsheremata/earthenterprise:
  Added test for test data. Commented out some code in order to compile.

+++ earth_enterprise/src/fusion/autoingest/sysman/InsetTilespaceIndex.cpp
@@ -35,7 +35,7 @@ Changes:
 #include <iostream>
 #include <stdint.h>
 #include <assert.h>
-
+/*
 QuadtreePath InsetTilespaceIndex::add(const khExtents <uint32> &extents) {
     int level;
     QuadtreePath quadtreeMbr = getQuadtreeMBR(extents, level, MAX_LEVEL);
@@ -55,8 +55,8 @@ QuadtreePath InsetTilespaceIndex::add(const khExtents <uint32> &extents) {
     mbrExtentsVec->push_back(&extents);
     return quadtreeMbr;
 }
-
-QuadtreePath InsetTilespaceIndex::getQuadtreeMBR(const khExtents<uint32>& extents, int& level, const int max_level) {
+*/
+QuadtreePath InsetTilespaceIndex::getQuadtreeMBR(const khExtents<double>& extents, int& level, const int max_level) {
     double north = 180;
     double south = -180;
     double west = -180;
